Reinstate include_gps options for special_locations
