+++ src/pint/observatory/special_locations.py
@@ -37,7 +37,7 @@ class SpecialLocation(Observatory):
     include_gps : bool, optional
         Set False to disable UTC(GPS)->UTC clock correction.
     include_bipm : bool, optional
-        Set False to disable UTC-> TT BIPM clock
+        Set False to disable TAI-> TT BIPM clock
         correction. If False, it only apply TAI->TT correction
         TT = TAI+32.184s, the same as TEMPO2 TT(TAI) in the
         parfile. If True, it will apply the correction from
@@ -45,9 +45,37 @@ class SpecialLocation(Observatory):
         https://www.bipm.org/en/bipm-services/timescales/time-ftp/ttbipm.html
     bipm_version : str, optional
         Set the version of TT BIPM clock correction file to
-        use, the default is %s.  It has to be in the format
-        like 'BIPM2015'
-    """ % bipm_default
+        use. It has to be in the format like 'BIPM2015'
+    overwrite : bool, optional
+        If True, allow redefinition of an existing observatory; if False,
+        raise an exception.
+    """
+
+    def __init__(
+        self,
+        name,
+        aliases=None,
+        include_gps=True,
+        include_bipm=True,
+        bipm_version=bipm_default,
+        overwrite=False,
+    ):
+        super().__init__(
+            name,
+            aliases=aliases,
+            include_gps=include_gps,
+            include_bipm=include_bipm,
+            bipm_version=bipm_version,
+            overwrite=overwrite,
+        )
+
+        self.origin = "Built-in special location."
+
+
+class BarycenterObs(SpecialLocation):
+    """Observatory-derived class for the solar system barycenter.
+
+    Time scale is assumed to be tdb."""
 
     def __init__(
         self,
@@ -55,25 +83,14 @@ class SpecialLocation(Observatory):
         aliases=None,
         overwrite=False,
     ):
-
-        self.origin = "Built-in special location."
-
-        super().__init__(name, aliases=aliases)
-
-    def clock_corrections(self, t, limits="warn"):
-        return np.zeros(t.shape) * u.s
-
-    def last_clock_correction_mjd(self):
-        """Return the MJD of the last available clock correction.
-
-        Returns ``np.inf`` if no clock corrections are relevant.
-        """
-        return np.inf
-
-
-class BarycenterObs(SpecialLocation):
-    """Observatory-derived class for the solar system barycenter.  Time
-    scale is assumed to be tdb."""
+        super().__init__(
+            name,
+            aliases=aliases,
+            include_gps=False,
+            include_bipm=False,
+            bipm_version=bipm_default,
+            overwrite=overwrite,
+        )
 
     @property
     def timescale(self):
@@ -228,4 +245,4 @@ class T2SpacecraftObs(SpecialLocation):
 BarycenterObs("barycenter", aliases=["@", "ssb", "bary", "bat"])
 GeocenterObs("geocenter", aliases=["0", "o", "coe", "geo"])
 T2SpacecraftObs("stl_geo", aliases=["STL_GEO"])
-# TODO -- BIPM issue
+# TODO -- how to handle user changes in BIPM choice?
