+++ src/pint/observatory/topo_obs.py
@@ -13,30 +13,22 @@ import numpy as np
 from astropy.coordinates import EarthLocation
 from loguru import logger as log
 
+import pint.observatory
 from pint import JD_MJD
-from pint.config import runtimefile
 from pint.erfautils import gcrs_posvel_from_itrf
 from pint.observatory import (
-    ClockCorrectionError,
-    ClockCorrectionOutOfRange,
     NoClockCorrections,
     Observatory,
     bipm_default,
     get_observatory,
+    find_clock_file,
 )
-from pint.observatory.clock_file import ClockFile, GlobalClockFile
-from pint.observatory.global_clock_corrections import Index, get_clock_correction_file
 from pint.pulsar_mjd import Time
 from pint.solar_system_ephemerides import get_tdb_tt_ephem_geocenter, objPosVel_wrt_SSB
 from pint.utils import has_astropy_unit
 
-pint_env_var = "PINT_CLOCK_OVERRIDE"
 
 __all__ = ["TopoObs", "find_clock_file", "export_all_clock_files"]
-
-# These are global because they are, well, literally global
-_gps_clock = None
-_bipm_clock_versions = {}
 
 
 class TopoObs(Observatory):
@@ -59,13 +51,12 @@ class TopoObs(Observatory):
 
     Parameters
     ----------
-
     name : str
         The name of the observatory
     itrf_xyz : astropy.units.Quantity or array-like
         IRTF site coordinates (len-3 array).  Can include
-       astropy units.  If no units are given, meters are
-       assumed.
+        astropy units.  If no units are given, meters are
+        assumed.
 
     tempo_code : str, optional
         1-character tempo code for the site.  Will be
@@ -88,7 +79,7 @@ class TopoObs(Observatory):
     include_gps : bool, optional
         Set False to disable UTC(GPS)->UTC clock correction.
     include_bipm : bool, optional
-        Set False to disable UTC-> TT BIPM clock
+        Set False to disable TAI-> TT BIPM clock
         correction. If False, it only apply TAI->TT correction
         TT = TAI+32.184s, the same as TEMPO2 TT(TAI) in the
         parfile. If True, it will apply the correction from
@@ -125,6 +116,20 @@ class TopoObs(Observatory):
         overwrite=False,
         bogus_last_correction=False,
     ):
+        if aliases is None:
+            aliases = []
+        for code in (tempo_code, itoa_code):
+            if code is not None:
+                aliases.append(code)
+
+        super().__init__(
+            name,
+            aliases=aliases,
+            include_gps=include_gps,
+            include_bipm=include_bipm,
+            bipm_version=bipm_version,
+            overwrite=overwrite,
+        )
         # ITRF coordinates are required
         if itrf_xyz is None:
             raise ValueError("ITRF coordinates not given for observatory '%s'" % name)
@@ -158,24 +163,10 @@ class TopoObs(Observatory):
         if clock_fmt == "tempo" and clock_file == "time.dat" and tempo_code is None:
             raise ValueError("No tempo_code set for observatory '%s'" % name)
 
-        # GPS corrections
-        self.include_gps = include_gps
-
-        # BIPM corrections
-        # WARNING: `get_observatory` changes these after construction
-        self.include_bipm = include_bipm
-        self.bipm_version = bipm_version
         self.bogus_last_correction = bogus_last_correction
 
         self.tempo_code = tempo_code
-        if aliases is None:
-            aliases = []
-        for code in (tempo_code, itoa_code):
-            if code is not None:
-                aliases.append(code)
-
         self.origin = origin
-        super().__init__(name, aliases=aliases)
 
     @property
     def timescale(self):
@@ -183,42 +174,6 @@ class TopoObs(Observatory):
 
     def earth_location_itrf(self, time=None):
         return self._loc_itrf
-
-    def _load_gps_clock(self):
-        global _gps_clock
-        if _gps_clock is None:
-            log.info(f"Loading global GPS clock file for {self.name}")
-            _gps_clock = find_clock_file(
-                "gps2utc.clk",
-                format="tempo2",
-                clock_dir=self.clock_dir,
-            )
-            if len(_gps_clock.time) == 0:
-                raise NoClockCorrections(
-                    "Unable to obtain GPS to UTC clock corrections"
-                )
-
-    def _load_bipm_clock(self, bipm_version):
-        bipm_version = bipm_version.lower()
-        if bipm_version not in _bipm_clock_versions:
-            try:
-                log.info(
-                    f"Loading BIPM clock version {bipm_version} " f"for {self.name}"
-                )
-                # FIXME: error handling?
-                _bipm_clock_versions[bipm_version] = find_clock_file(
-                    f"tai2tt_{bipm_version}.clk",
-                    format="tempo2",
-                    clock_dir=self.clock_dir,
-                )
-            except Exception as e:
-                raise ValueError(
-                    f"Cannot find TT BIPM file for version '{bipm_version}'."
-                ) from e
-            if len(_bipm_clock_versions[bipm_version].time) == 0:
-                raise NoClockCorrections(
-                    f"Unable to obtain BIPM {bipm_version} clock corrections"
-                )
 
     def _load_clock_corrections(self):
         if self._clock is None:
@@ -247,8 +202,8 @@ class TopoObs(Observatory):
             The time when the clock correcions are applied.
         """
 
+        corr = super().clock_corrections(t, limits=limits)
         # Read clock file if necessary
-        corr = np.zeros_like(t) * u.us
         self._load_clock_corrections()
         if not self._clock:
             if self.clock_file:
@@ -265,23 +220,6 @@ class TopoObs(Observatory):
             for clock in self._clock:
                 corr += clock.evaluate(t, limits=limits)
 
-        if self.include_gps:
-            log.info("Applying GPS to UTC clock correction (~few nanoseconds)")
-            self._load_gps_clock()
-            corr += _gps_clock.evaluate(t, limits=limits)
-
-        if self.include_bipm:
-            log.info(
-                f"Applying TT(TAI) to TT({self.bipm_version}) clock correction (~27 us)"
-            )
-            tt2tai = 32.184 * 1e6 * u.us
-            self._load_bipm_clock(self.bipm_version)
-            corr += (
-                _bipm_clock_versions[self.bipm_version.lower()].evaluate(
-                    t, limits=limits
-                )
-                - tt2tai
-            )
         return corr
 
     def last_clock_correction_mjd(self):
@@ -290,20 +228,10 @@ class TopoObs(Observatory):
         Combines constraints based on Earth orientation parameters and on the
         available clock corrections specific to the telescope.
         """
-        t = np.inf
+        t = super().last_clock_correction_mjd()
         self._load_clock_corrections()
-        if not self._clock:
-            return -np.inf
         for clock in self._clock:
             t = min(t, clock.last_correction_mjd())
-        if self.include_gps:
-            self._load_gps_clock()
-            t = min(t, _gps_clock.last_correction_mjd())
-        if self.include_bipm:
-            self._load_bipm_clock(self.bipm_version)
-            t = min(
-                t, _bipm_clock_versions[self.bipm_version.lower()].last_correction_mjd()
-            )
         return t
 
     def _get_TDB_ephem(self, t, ephem):
@@ -363,131 +291,6 @@ class TopoObs(Observatory):
         return obs_geocenter_pv + earth_pv
 
 
-# FIXME: is clock_file the right place for this? The rest of the file isn't very
-# PINT-specific but this is. Maybe in topo_obs - does that cover all the clock files
-# we need?
-def find_clock_file(
-    name, format, bogus_last_correction=False, url_base=None, clock_dir=None
-):
-    """Locate and return a ClockFile in one of several places.
-
-    PINT looks for clock files in three places, in order:
-
-    1. The directory ``$PINT_CLOCK_OVERRIDE``
-    2. The global clock correction repository on the Internet (or a locally cached copy)
-    3. The directory ``pint.config.runtimefile('.')``
-
-    The first place the file is found is the one use; this allows you to force PINT to
-    use your own files in place of those in the global repository.
-
-    Parameters
-    ----------
-    name : str
-        The name of the file, for example ``time_ao.dat``.
-    format : "tempo" or "tempo2"
-        The format of the file; this also determines where in the global repository
-        to look for it.
-    bogus_last_correction : bool
-        Whether the file contains a far-future value to help other programs'
-        interpolation cope.
-    url_base : str or None
-        Override the usual location to look for global clock corrections
-        (mostly useful for testing)
-    clock_dir : str or pathlib.Path or None
-        If None or "PINT", use the above procedure; if "TEMPO" or "TEMPO2" use
-        those programs' customary locations; if a path, look there specifically.
-
-    Returns
-    -------
-    ClockFile
-    """
-    if name == "":
-        raise ValueError("No filename supplied to find_clock_file")
-    if clock_dir is None or str(clock_dir).upper() == "PINT":
-        # Don't try loading it from a specific path
-        p = None
-    elif str(clock_dir).lower() == "tempo":
-        if "TEMPO" not in os.environ:
-            raise NoClockCorrections(
-                f"TEMPO environment variable not set but clock file {name} "
-                f"is supposed to be in the directory it points to"
-            )
-        p = Path(os.environ["TEMPO"]) / "clock" / name
-    elif str(clock_dir).lower() == "tempo2":
-        if "TEMPO2" not in os.environ:
-            raise NoClockCorrections(
-                f"TEMPO2 environment variable not set but clock file {name} "
-                f"is supposed to be in the directory it points to"
-            )
-        # Look in the TEMPO2 directory and nowhere else
-        p = Path(os.environ["TEMPO2"]) / "clock" / name
-    else:
-        # assume it's a path and look in there
-        p = Path(clock_dir) / name
-    if p is not None:
-        log.info("Loading clock file {p} from specified location")
-        return ClockFile.read(
-            p,
-            format=format,
-            bogus_last_correction=bogus_last_correction,
-            friendly_name=name,
-        )
-
-    env_clock = None
-    global_clock = None
-    local_clock = None
-    if pint_env_var in os.environ:
-        loc = Path(os.environ[pint_env_var]) / name
-        if loc.exists():
-            # FIXME: more arguments?
-            env_clock = ClockFile.read(
-                loc,
-                format=format,
-                bogus_last_correction=bogus_last_correction,
-                friendly_name=name,
-            )
-            # Could just return this but we want to emit
-            # a warning with an appropriate level of forcefulness
-    index = Index(url_base=url_base)
-    if name in index.files:
-        global_clock = GlobalClockFile(
-            name,
-            format=format,
-            bogus_last_correction=bogus_last_correction,
-            url_base=url_base,
-        )
-    loc = Path(runtimefile(name))
-    if loc.exists():
-        local_clock = ClockFile.read(
-            loc,
-            format=format,
-            bogus_last_correction=bogus_last_correction,
-            friendly_name=name,
-        )
-
-    if env_clock is not None:
-        if global_clock is not None:
-            # FIXME: if we're not going to use the values from the global clock
-            # we could have saved downloading and parsing it
-            log.warning(
-                f"Clock file from {env_clock.filename} overrides global clock "
-                f"file {name} because of {pint_env_var}"
-            )
-        else:
-            log.info(f"Using clock file from {env_clock.filename}")
-        return env_clock
-    elif global_clock is not None:
-        log.info(f"Using global clock file for {name} with {bogus_last_correction=}")
-        return global_clock
-    elif local_clock is not None:
-        log.info(f"Using local clock file for {name}")
-        return local_clock
-    else:
-        # Null clock file should return warnings/exceptions if ever you try to
-        # look up a data point in it
-        raise NoClockCorrections(f"No clock file for {name}")
-
-
 def export_all_clock_files(directory):
     """Export all clock files PINT is using.
 
@@ -506,9 +309,11 @@ def export_all_clock_files(directory):
         Where to put the files.
     """
     directory = Path(directory)
-    if _gps_clock is not None:
-        _gps_clock.export(directory / Path(_gps_clock.filename).name)
-    for version, clock in _bipm_clock_versions.items():
+    if pint.observatory._gps_clock is not None:
+        pint.observatory._gps_clock.export(
+            directory / Path(pint.observatory._gps_clock.filename).name
+        )
+    for version, clock in pint.observatory._bipm_clock_versions.items():
         clock.export(directory / Path(clock.filename).name)
     for name in Observatory.names():
         o = get_observatory(name)
